Lab1 task 2 added

+++ Lab1/task1/src/Caesar.java
@@ -1,6 +1,6 @@
 public class Caesar {
 
-    public static String decodeXor(String in, int key) {
+    public static String decodeXor(String in, byte key) {
         StringBuilder out = new StringBuilder();
 
         for (int i = 0; i < in.length(); i++) {
